requirements.txt
asttokens==3.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
easyocr==1.7.2
executing==2.2.0
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.3.0
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.1.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
Pygments==2.19.1
pyparsing==3.2.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytubefix==8.12.2
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
requests==2.32.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==76.0.0
shapely==2.0.7
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.6.0
tifffile==2025.3.13
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.88
ultralytics-thop==2.0.14
urllib3==2.3.0
wcwidth==0.2.13
